requirements.txt
hydra-core==1.1.1
numpy==1.22.0
omegaconf==2.1.1
torch==1.10.0
